pyproject.toml
[project]
name = "aqi-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.8",
    "hopsworks[python]==4.2.*",
    "ipykernel>=7.1.0",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numba>=0.61.0",
    "numpy<2.0",
    "pandas<2.2.0",
    "prophet>=1.2.1",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "fastapi>=0.109.0",
    "idna<3.7", # Pin idna to avoid conflicts if needed, or let resolver handle it. kept safe
    "plotly>=5.18.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "streamlit>=1.30.0",
    "tensorflow>=2.15.0",
    "uvicorn>=0.27.0",
    "xgboost==1.7.6",
]
